Move Schema extensions in sub namespace
